OOPM JAVA: Rename Practical 5 to 6
